Safari-611.1.10/LayoutTests/http/tests/xmlhttprequest/workers/access-control-basic-get-fail-non-simple-expected.txt

CONSOLE MESSAGE: Preflight response is not successful
GET should not trigger a preflight request from a worker unless it has non-simple headers.

PASS: Cross-domain access allowed for simple get.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [ A network error occurred.].
DONE
